batch2b: fix C8/C10/C11/M10 audit issues
- C8: AABB perf — precompute local AABB (min/max), transform 8 corners
- C10: RingBuffer SPSC memory ordering — derive read_index from write_index, fences
- C11: GPU test memory leak — DeviceMem RAII guard
- M10: render() silent no-op -> throw std::runtime_error
- PLAN.md: add Batch Cleanup section

+++ libs/ure_sceneio/src/image_loader.cpp
@@ -21,7 +21,7 @@ bool load_image_rgb32f(const std::string& file_path, gpu::HostTexture& out_textu
 
     out_texture.width = w;
     out_texture.height = h;
-    out_texture.data.assign(data, data + static_cast<size_t>(w) * h * 3);
+    out_texture.data.assign(data, data + static_cast<size_t>(w) * static_cast<size_t>(h) * 3);
     stbi_image_free(data);
     return true;
 }
